@@ -6,8 +6,8 @@ import android.app.Activity;
 import android.content.Context;
 import android.content.ContextWrapper;
 import android.os.Bundle;
-import android.support.annotation.IntegerRes;
-import android.support.annotation.NonNull;
+import androidx.annotation.IntegerRes;
+import androidx.annotation.NonNull;
 import android.support.v4.media.MediaBrowserCompat.MediaItem;
 import android.view.View;
 import android.view.ViewPropertyAnimator;
