@@ -1,76 +1,43 @@
 package io.straas.android.media.demo.widget;
 
-import android.content.Context;
-import android.content.res.TypedArray;
-import android.graphics.drawable.Drawable;
-import android.graphics.drawable.GradientDrawable;
-import android.os.Build;
-import android.os.Bundle;
-import android.support.annotation.DrawableRes;
-import android.support.annotation.IntDef;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.annotation.StyleRes;
-import android.support.graphics.drawable.VectorDrawableCompat;
-import android.support.v4.app.FragmentActivity;
-import android.support.v4.content.ContextCompat;
-import android.support.v4.media.MediaBrowserCompat.ConnectionCallback;
-import android.support.v4.media.MediaMetadataCompat;
-import android.support.v4.media.session.MediaControllerCompat;
-import android.support.v4.media.session.MediaSessionCompat.QueueItem;
-import android.support.v4.media.session.PlaybackStateCompat;
-import android.support.v4.util.SparseArrayCompat;
-import android.support.v4.view.GestureDetectorCompat;
-import android.support.v4.widget.TextViewCompat;
-import android.support.v7.widget.AppCompatImageButton;
-import android.support.v7.widget.AppCompatImageView;
-import android.text.SpannableStringBuilder;
-import android.text.TextUtils;
-import android.util.AttributeSet;
-import android.util.Log;
-import android.view.GestureDetector;
-import android.view.LayoutInflater;
-import android.view.MotionEvent;
-import android.view.View;
-import android.view.ViewGroup;
-import android.view.ViewParent;
-import android.widget.FrameLayout;
-import android.widget.ImageButton;
-import android.widget.ImageView;
-import android.widget.ProgressBar;
-import android.widget.RelativeLayout;
-import android.widget.TextView;
+import android.content.*;
+import android.content.res.*;
+import android.graphics.drawable.*;
+import android.os.*;
+import android.support.v4.media.MediaBrowserCompat.*;
+import android.support.v4.media.*;
+import android.support.v4.media.session.*;
+import android.support.v4.media.session.MediaSessionCompat.*;
+import android.text.*;
+import android.util.*;
+import android.view.*;
+import android.widget.*;
 
-import com.bumptech.glide.Glide;
-import com.bumptech.glide.request.RequestOptions;
-import com.google.android.exoplayer2.Format;
+import com.bumptech.glide.*;
+import com.bumptech.glide.request.*;
+import com.google.android.exoplayer2.*;
 
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
+import java.lang.annotation.*;
+import java.util.*;
 
-import io.straas.android.media.demo.MediaControllerCompatHelper;
-import io.straas.android.media.demo.MediaControllerCompatHelper.VideoQualityInfo;
-import io.straas.android.media.demo.MediaControllerCompatHelper.VideoQualityInfoCallback;
+import androidx.annotation.*;
+import androidx.appcompat.widget.*;
+import androidx.collection.*;
+import androidx.core.content.*;
+import androidx.core.view.*;
+import androidx.core.widget.*;
+import androidx.fragment.app.*;
+import androidx.vectordrawable.graphics.drawable.*;
+import io.straas.android.media.demo.*;
 import io.straas.android.media.demo.R;
 import io.straas.android.media.demo.Utils;
-import io.straas.android.media.demo.widget.ui.ContentSeekBar;
-import io.straas.android.media.demo.widget.ui.SwitchQualityDialog;
-import io.straas.android.media.demo.widget.ui.SwitchSpeedDialog;
-import io.straas.android.sdk.media.StraasMediaCore;
-import io.straas.android.sdk.media.StraasMediaCore.ErrorReason;
-import io.straas.android.sdk.media.VideoCustomMetadata;
+import io.straas.android.media.demo.MediaControllerCompatHelper.*;
+import io.straas.android.media.demo.widget.ui.*;
+import io.straas.android.sdk.media.*;
+import io.straas.android.sdk.media.StraasMediaCore.*;
 
-import static io.straas.android.sdk.media.LiveEventListener.BROADCAST_STATE_DVR_PLAYBACK_AVAILABLE;
-import static io.straas.android.sdk.media.LiveEventListener.BROADCAST_STATE_ENDED;
-import static io.straas.android.sdk.media.LiveEventListener.BROADCAST_STATE_STARTED;
-import static io.straas.android.sdk.media.LiveEventListener.BROADCAST_STATE_STOPPED;
-import static io.straas.android.sdk.media.LiveEventListener.BROADCAST_STATE_UNKNOWN;
-import static io.straas.android.sdk.media.LiveEventListener.BROADCAST_STATE_WAITING_FOR_STREAM;
-import static io.straas.android.sdk.media.VideoCustomMetadata.KEY_TEXT_TRACKS;
-import static io.straas.android.sdk.media.VideoCustomMetadata.LIVE_BROADCAST_STATE_V2;
+import static io.straas.android.sdk.media.LiveEventListener.*;
+import static io.straas.android.sdk.media.VideoCustomMetadata.*;
 
 public final class StraasPlayerView extends FrameLayout implements StraasMediaCore.UiContainer {
     private static final String TAG = StraasPlayerView.class.getSimpleName();
