@@ -2,12 +2,12 @@ package io.straas.android.sdk.streaming.demo.camera;
 
 import android.content.*;
 import android.os.*;
-import android.support.annotation.*;
-import android.support.v7.app.*;
 import android.text.*;
 import android.view.*;
 import android.widget.*;
 
+import androidx.annotation.*;
+import androidx.appcompat.app.*;
 import io.straas.android.sdk.demo.common.widget.*;
 import io.straas.android.sdk.streaming.demo.*;
 
