@@ -2,13 +2,9 @@ package io.straas.android.sdk.streaming.demo.camera;
 
 import android.*;
 import android.app.*;
-import android.arch.lifecycle.Observer;
 import android.content.*;
 import android.content.pm.*;
 import android.os.*;
-import android.support.annotation.*;
-import android.support.v4.app.*;
-import android.support.v7.app.*;
 import android.text.*;
 import android.util.*;
 import android.view.*;
@@ -18,6 +14,10 @@ import com.google.android.gms.tasks.*;
 
 import java.util.*;
 
+import androidx.annotation.*;
+import androidx.appcompat.app.*;
+import androidx.core.app.*;
+import androidx.lifecycle.Observer;
 import io.straas.android.sdk.demo.common.*;
 import io.straas.android.sdk.demo.qrcode.*;
 import io.straas.android.sdk.media.*;
